+++ src/libponyc/pass/names.c
@@ -61,7 +61,7 @@ static bool names_applycap(ast_t* ast, ast_t* cap, ast_t* ephemeral)
 
 static bool names_resolvealias(ast_t* def, ast_t* type)
 {
-  ast_state_t state = (ast_state_t)((uint32_t)ast_data(def));
+  ast_state_t state = (ast_state_t)((uint64_t)ast_data(def));
 
   switch(state)
   {
